refactor(app): Implementa suporte a valores decimais no estoque e logs
Esta alteração introduz o uso de BigDecimal para quantidades de estoque e histórico, permitindo o controle de produtos fracionados (ex: m² de vidro).

## Alterações Principais:

* **Estoque:** Refatora campos de quantidade (`quantidade_total`, `quantidade_disponivel`, `reservado`) para usar `DECIMAL(18, 5)` no banco e `BigDecimal` nas entidades Java.
* **Item Pedido:** Implementa a lógica em memória para calcular e retornar o `subtotal` corretamente no DTO, resolvendo o problema de colunas geradas (`insertable=false`).
* **Logs/Histórico:** Corrige erros de formatação (`IllegalFormatConversionException`) ao usar `%s` em `String.format()` para garantir a exibição correta dos valores `BigDecimal` no histórico de movimentação.

+++ src/main/java/com/project/extension/dto/estoque/EstoqueResponseDto.java
@@ -2,11 +2,13 @@ package com.project.extension.dto.estoque;
 
 import com.project.extension.dto.produto.ProdutoResponseDto;
 
+import java.math.BigDecimal;
+
 public record EstoqueResponseDto(
         Integer id,
-        Integer quantidadeTotal,
-        Integer quantidadeDisponivel,
-        Integer reservado,
+        BigDecimal quantidadeTotal,
+        BigDecimal quantidadeDisponivel,
+        BigDecimal reservado,
         String localizacao,
         ProdutoResponseDto produto
 ) {
