+++ src/main/java/com/project/extension/dto/historicoestoque/HistoricoEstoqueResponseDto.java
@@ -4,13 +4,14 @@ import com.project.extension.dto.estoque.EstoqueResponseDto;
 import com.project.extension.dto.usuario.UsuarioResponseDto;
 import com.project.extension.entity.TipoMovimentacao;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record HistoricoEstoqueResponseDto(
         Integer id,
         TipoMovimentacao tipoMovimentacao,
-        Integer quantidade,
-        Integer quantidadeAtual,
+        BigDecimal quantidade,
+        BigDecimal quantidadeAtual,
         String observacao,
         LocalDateTime dataHora,
         EstoqueResponseDto estoque,
